Added objective_function.py and sogtest
Abstract base class + example of how to use it. Not too fond of the name
nor the setup. I'm thinking writing the input file would be the move to
start with because the point of objective_function.py is that you could
just specify custom code but you don't have to have the python file
every time you want to optimize something. That's not in line with the
project.